Updated the repos.
The resource is a roadmap to help with SQL for Data Science Interviews. This roadmap provides an overview of the topics related to SQL for Data Science Interviews, including the topics that are commonly covered, the key concepts and techniques to master, and example questions. This resource should be of great value to the learners if preparing for upcoming interviews.